refactor(HttpMethod): HttpMethod 객체를 request 패키지로 이동
