+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -4,7 +4,6 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.Map;
 import java.util.Optional;
-import org.apache.coyote.http11.HttpMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
